updated lodview_util (important change)
