+++ src/util/lodview_util.py
@@ -1,17 +1,16 @@
 import logging
 import json
 from flask import render_template, Response
-from rdflib import Graph, URIRef, Literal, BNode, Namespace
-from rdflib.namespace import RDF, RDFS, SDO, SKOS, DCTERMS  # type: ignore
+from rdflib import Graph, URIRef, Literal, BNode
+from rdflib.namespace import RDF, RDFS, SDO, SKOS, DCTERMS, OWL  # type:ignore
 from typing import Optional, List
 from util.mime_type_util import MimeType
 from util.APIUtil import APIUtil
 import util.mw_ld_util
+from util.ns_util import SCHEMA, MUZIEKWEB_VOCAB, SKOSXL
+from config import cfg
 
 logger = logging.getLogger()
-
-SKOSXL_NS = "http://www.w3.org/2008/05/skos-xl#"
-SCHEMA = Namespace("http://schema.org/")  # Note the difference with SDO above (http:).
 
 
 def json_ld_structured_data_for_resource(rdf_graph: Graph, resource_url: str) -> str:
@@ -29,43 +28,132 @@ def json_ld_structured_data_for_resource(rdf_graph: Graph, resource_url: str) ->
 # ========= JSON generator functions for lod-view ==========
 
 
+def json_parts_from_IRI(rdf_graph: Graph, iri: str) -> dict:
+    """Generates the parts needed in the UI, eg. prefix, namespace and property."""
+    if iri.endswith("/"):  # URIs that end with a '/' can not be split by rdflib.
+        return {"uri": iri, "prefix": "", "namespace": "", "property": ""}
+    else:
+        prefix, namespace, name = rdf_graph.compute_qname(str(iri))
+        return {
+            "uri": iri,
+            "prefix": prefix,
+            "namespace": str(namespace),
+            "property": name,
+        }
+
+
+def json_label_for_node(
+    rdf_graph: Graph, node: URIRef | BNode, lang: str = ""
+) -> List[str]:
+    """Generates a list of all possible labels for a given IRI or BNode.
+    This list is used for the visualisation of the resource in the LOD view.
+    :param lang: the language preference for the labels.
+    """
+    my_literal_list = (
+        [Literal(label) for label in rdf_graph.objects(node, SKOS.prefLabel)]
+        + [Literal(name) for name in rdf_graph.objects(node, SDO.name)]
+        + [Literal(title) for title in rdf_graph.objects(node, DCTERMS.title)]
+        + [Literal(label) for label in rdf_graph.objects(node, RDFS.label)]
+        + [
+            Literal(lf)
+            for lf in rdf_graph.objects(node, URIRef(f"{SKOSXL}literalForm"))
+        ]
+    )
+    return [get_string_for_langstring(label, lang) for label in my_literal_list]
+
+
+def get_string_for_langstring(literal: Literal, lang: str = "") -> str:
+    """Returns the string value of a Literal, taking into account the
+    language preference. If the literal has a language tag that matches
+    the preferred language, the string value is returned without the language
+    tag. If the literal has a language tag that does not match the preferred
+    language, the string value is returned with the language tag. For other
+    cases (e.g. no language tag defined), the string value is returned without
+    the language tag.
+    :param lang: the language preference for the labels.
+    """
+    if literal.language is None:
+        # no language tag defined: return the string value
+        return str(literal)
+    elif literal.language and lang == "":
+        # no preferred language defined but language tag defined:
+        # return the string value including language tag, to make it clear that
+        # there is a language tag defined
+        return f"{str(literal)} @{literal.language}"
+    elif literal.language and lang != "" and literal.language == lang:
+        # language tag defined, there is a preferred language and it matches the
+        # literal's language: return the string value (because the UI is all
+        # in that preferred language, no need to add the language tag)
+        return str(literal)
+    elif literal.language and literal.language != lang:
+        # language tag defined but does not match the preferred language:
+        # return the string value with the language tag
+        return f"{str(literal)} @{literal.language}"
+    else:
+        # this should not happen, but in case it does, return the empty string
+        return ""
+
+
+def json_for_literal(rdf_graph: Graph, literal: Literal, lang: str = "") -> dict:
+    """Generates JSON parts for a given Literal."""
+    return json_label_for_literal(literal, lang) | json_for_datatype(
+        rdf_graph, literal.datatype
+    )
+
+
+def json_for_datatype(rdf_graph: Graph, datatype_iri: URIRef | None) -> dict:
+    """Generates JSON parts for a given datatype IRI."""
+    if datatype_iri:
+        prefix, namespace, name = rdf_graph.compute_qname(str(datatype_iri))
+        return {
+            "datatype": {
+                "uri": str(datatype_iri),
+                "prefix": prefix,
+                "namespace": str(namespace),
+                "property": name,
+            }
+        }
+    else:
+        return {}
+
+
+def json_label_for_literal(literal: Literal, lang: str = "") -> dict:
+    """Generates a list of labels for a given Literal, taking language
+    preference into account.
+    :param lang: the language preference for the labels.
+    """
+    return {"labels": [get_string_for_langstring(literal, lang)]}
+    # | json_for_datatype(literal.graph, literal.datatype)
+
+
 def json_header_from_rdf_graph(
     rdf_graph: Graph, resource_url: str
 ) -> Optional[List[dict]]:
-    """Generates the data for the header in resource.html"""
+    """Generates the JSON data for the header in lodview."""
     json_header = []
     try:
         json_header = [
             {
-                "title": [
-                    str(label)
-                    for label in rdf_graph.objects(URIRef(resource_url), SKOS.prefLabel)
-                    if Literal(label).language == "nl"
-                ]
-                + [
-                    str(name)
-                    for name in rdf_graph.objects(URIRef(resource_url), SDO.name)
-                ]
-                + [
-                    str(name)
-                    for name in rdf_graph.objects(URIRef(resource_url), DCTERMS.title)
-                ]
-                + [
-                    str(label)
-                    for label in rdf_graph.objects(URIRef(resource_url), RDFS.label)
+                "title": json_label_for_node(
+                    rdf_graph,
+                    URIRef(resource_url),
+                    lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl"),
+                ),
+                "o": [
+                    json_parts_from_IRI(rdf_graph, str(o))
+                    for o in rdf_graph.objects(
+                        subject=URIRef(resource_url), predicate=URIRef(RDF.type)
+                    )
+                    if json_parts_from_IRI(rdf_graph, str(o)).get("namespace", "")
+                    in (
+                        str(SDO),
+                        str(SKOS),
+                        str(SCHEMA),
+                        str(MUZIEKWEB_VOCAB),
+                        str(OWL),
+                    )
                 ],
-                "o": {
-                    "uri": str(o),
-                    "prefix": rdf_graph.compute_qname(str(o))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(o))[1]),
-                    "property": rdf_graph.compute_qname(str(o))[2],
-                },
             }
-            for o in rdf_graph.objects(
-                subject=URIRef(resource_url), predicate=URIRef(RDF.type)
-            )
-            if str(rdf_graph.compute_qname(str(o))[1])
-            in (str(SDO), str(SKOS), str(SCHEMA))
         ]
         return json_header
     except Exception as e:
@@ -77,65 +165,22 @@ def json_header_from_rdf_graph(
 def json_iri_iri_from_rdf_graph(
     rdf_graph: Graph, resource_url: str
 ) -> Optional[List[dict]]:
-    """Generates part of the LOD view data for resource.html that with: (<uri> <uri> <uri>)"""
+    """Generates JSON data (<uri> <uri> <uri>) for the LOD view."""
     json_iri_iri = []
     try:
         json_iri_iri = [
             {
-                "p": {
-                    "uri": str(p),
-                    "prefix": rdf_graph.compute_qname(str(p))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                    "property": rdf_graph.compute_qname(str(p))[2],
-                },
-                "o": {
-                    "uri": str(o),
-                    "literal_form": [
-                        f"{str(lf)} @{Literal(lf).language}"
-                        for lf in rdf_graph.objects(
-                            o, URIRef(f"{SKOSXL_NS}literalForm")
-                        )
-                    ],
-                    "prefix": rdf_graph.compute_qname(str(o))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(o))[1]),
-                    "property": rdf_graph.compute_qname(str(o))[2],
-                    "pref_label": [
-                        str(label)
-                        for label in rdf_graph.objects(o, SKOS.prefLabel)
-                        if Literal(label).language == "nl"
-                    ],
-                    "parent_label": [
-                        str(label) for label in rdf_graph.objects(o, SDO.name)
-                    ],
-                    "part_label": [
-                        str(label) for label in rdf_graph.objects(o, SDO.name)
-                    ],
+                "p": json_parts_from_IRI(rdf_graph, str(p)),
+                "o": json_parts_from_IRI(rdf_graph, str(o))
+                | {
+                    "labels": json_label_for_node(
+                        rdf_graph, o, lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl")
+                    )
                 },
             }
             for (p, o) in rdf_graph.predicate_objects(subject=URIRef(resource_url))
-            if p != RDF.type and isinstance(o, URIRef) and not str(o).endswith("/")
+            if isinstance(o, URIRef)
         ]
-        # URIs that end with a '/' can not be split by rdflib. Therefore we add them separately.
-        for p, o in rdf_graph.predicate_objects(subject=URIRef(resource_url)):
-            if isinstance(o, URIRef) and str(o).endswith("/"):
-                json_iri_iri += [
-                    {
-                        "p": {
-                            "uri": str(p),
-                            "prefix": rdf_graph.compute_qname(str(p))[0],
-                            "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                            "property": rdf_graph.compute_qname(str(p))[2],
-                        },
-                        "o": {
-                            "uri": str(o),
-                            "literal_form": [],
-                            "prefix": "",
-                            "namespace": "",
-                            "property": "",
-                            "pref_label": [],
-                        },
-                    }
-                ]
         return json_iri_iri
     except Exception as e:
         logger.exception(f"Error in json_iri_iri_from_rdf_graph: {str(e)}")
@@ -146,38 +191,23 @@ def json_iri_iri_from_rdf_graph(
 def json_iri_lit_from_rdf_graph(
     rdf_graph: Graph, resource_url: str
 ) -> Optional[List[dict]]:
-    """Generates part of the LOD view data for resource.html (<uri> <uri> literal)"""
+    """Generates JSON structured data (<uri> <uri> literal) to be used in the LOD view templates."""
     json_iri_lit = []
     try:
         json_iri_lit = [
             {
-                "p": {
-                    "uri": str(p),
-                    "prefix": rdf_graph.compute_qname(str(p))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                    "property": rdf_graph.compute_qname(str(p))[2],
-                },
-                "o": {
-                    "literal_value": (
-                        f"{str(o)} @{o.language}" if o.language else f"{str(o)}"
-                    ),
-                    "datatype": str(o.datatype) if o.datatype is not None else "",
-                    "datatype_prefix": (
-                        rdf_graph.compute_qname(str(o.datatype))[0]
-                        if o.datatype is not None
-                        else ""
-                    ),
-                    "datatype_namespace": (
-                        str(rdf_graph.compute_qname(str(o.datatype))[1])
-                        if o.datatype is not None
-                        else ""
-                    ),
-                    "datatype_property": (
-                        rdf_graph.compute_qname(str(o.datatype))[2]
-                        if o.datatype is not None
-                        else ""
-                    ),
-                },
+                "p": json_parts_from_IRI(rdf_graph, str(p)),
+                "o": (
+                    json_label_for_literal(
+                        o,
+                        lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl"),
+                    )
+                    if o.datatype is None
+                    else json_label_for_literal(
+                        o, lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl")
+                    )
+                    | json_for_datatype(rdf_graph, o.datatype)
+                ),
             }
             for p, o in rdf_graph.predicate_objects(subject=URIRef(resource_url))
             if isinstance(o, Literal)
@@ -200,38 +230,23 @@ def json_iri_bnode_from_rdf_graph(
             if p != RDF.type and isinstance(o, BNode):
                 bnode_content = [
                     {
-                        "pred": {
-                            "uri": str(bnode_pred),
-                            "prefix": rdf_graph.compute_qname(str(bnode_pred))[0],
-                            "namespace": str(
-                                rdf_graph.compute_qname(str(bnode_pred))[1]
-                            ),
-                            "property": rdf_graph.compute_qname(str(bnode_pred))[2],
-                        },
+                        "pred": json_parts_from_IRI(rdf_graph, str(bnode_pred)),
                         "obj": (
-                            {
-                                "uri": str(bnode_obj),
-                                "prefix": rdf_graph.compute_qname(str(bnode_obj))[0],
-                                "namespace": str(
-                                    rdf_graph.compute_qname(str(bnode_obj))[1]
-                                ),
-                                "property": rdf_graph.compute_qname(str(bnode_obj))[2],
-                                "pref_label": [
-                                    f"{str(pl)} @{Literal(pl).language}"
-                                    for pl in rdf_graph.objects(
-                                        bnode_obj, SKOS.prefLabel
-                                    )
-                                ],
+                            json_parts_from_IRI(rdf_graph, str(bnode_obj))
+                            | {
+                                "labels": json_label_for_node(
+                                    rdf_graph,
+                                    bnode_obj,
+                                    lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl"),
+                                )
                             }
                             if isinstance(bnode_obj, URIRef)
                             else (
-                                {
-                                    "label": (
-                                        f"{str(bnode_obj)} @{bnode_obj.language}"
-                                        if bnode_obj.language
-                                        else f"{str(bnode_obj)}"
-                                    )
-                                }
+                                json_for_literal(
+                                    rdf_graph,
+                                    bnode_obj,
+                                    lang=cfg.get("UI_LANGUAGE_PREFERENCE", "nl"),
+                                )
                                 if isinstance(bnode_obj, Literal)
                                 else str(bnode_obj)
                             )
@@ -245,12 +260,7 @@ def json_iri_bnode_from_rdf_graph(
                 ]
                 json_iri_bnode.append(
                     {
-                        "p": {
-                            "uri": str(p),
-                            "prefix": rdf_graph.compute_qname(str(p))[0],
-                            "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                            "property": rdf_graph.compute_qname(str(p))[2],
-                        },
+                        "p": json_parts_from_IRI(rdf_graph, str(p)),
                         "o": bnode_content,
                     }
                 )
@@ -315,6 +325,7 @@ def get_lod_view_resource(
             album_art=util.mw_ld_util.get_album_art_from_rdf_graph(
                 rdf_graph, resource_url
             ),
+            pref_language=cfg.get("UI_LANGUAGE_PREFERENCE", "nl").upper(),
         )
     return ""
 
@@ -340,5 +351,5 @@ def get_lod_view_resource_header(
     return render_template(
         "resource_lod_view_header.html",
         resource_uri=resource_uri,
-        json_header=rdf_graph_header_json,
+        rdf_graph_header_json=rdf_graph_header_json,
     )
